Add offline first for Country
Adjust SwiftData conformance to Country
Properties on Country that cannot be serialised in SwiftData
have been converted to JSON strings that are decoded on the fly.

Add LocalCountriesProvider modelActor that loads, fetches and updates
locally stored countries.
For now only isFavorite flag is updated

In CountrySearchViewModel countries are loaded from db first, if there
are no locally stored countries the API used.